Remove new test staging data
Turns out the HTMLFileSetServ code handles spaces in urls just like any
other character, so no need for a special test case.

The Servlet API returns a 400 when there are *unencoded* spaces in the
url, but works fine with encoded spaces. If there are unencoded url
spaces, the HTMLFileSetServ code isn't called at all.

Keith Keller has confirmed that if you curl the HTMLFileSetServ dynamic
service container on CI with encoded spaces in the url everything works,
but if you go through nginx if barfs, so the problem is almost certainly
in the nginx url rewriting rules.

Keeping the other test cleanup though.